Removed Cached<Matrix4> from Transform

+++ scripts/clean.sh
@@ -9,9 +9,11 @@ BeginHeader "CLEANING"
 
 echo "Removing build${RELEASE_OR_DEBUG}..."
 rm -rf build${RELEASE_OR_DEBUG}
+if [ $? != 0 ] ; then Error "ERROR" ; exit 1 ; fi
 
 echo "Removing bin/${RELEASE_OR_DEBUG}..."
 rm -rf bin/${RELEASE_OR_DEBUG}
+if [ $? != 0 ] ; then Error "ERROR" ; exit 1 ; fi
 
 Success "Successfully cleaned!"
